# services/auth/api/docs/swagger.yaml
definitions:
  github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_api_models.RegisterOtp:
    properties:
      data:
        additionalProperties: {}
        type: object
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_api_models.Verify:
    properties:
      apple_code:
        type: string
      data:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_object_builder_service.V2LoginResponse'
      google_token:
        type: string
      otp:
        type: string
      provider:
        type: string
      register_type:
        type: string
      session_info:
        type: string
      tables:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Object'
        type: array
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.ClientPlatform:
    properties:
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      subdomain:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Company:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      logo:
        type: string
      name:
        type: string
      owner_id:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.LogoutRequest:
    properties:
      access_token:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Object:
    properties:
      object_id:
        type: string
      table_slug:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Role:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Session:
    properties:
      client_id:
        type: string
      client_platform_id:
        type: string
      client_type_id:
        type: string
      created_at:
        type: string
      data:
        type: string
      env_id:
        type: string
      expires_at:
        type: string
      id:
        type: string
      integration_id:
        type: string
      ip:
        type: string
      is_changed:
        type: boolean
      is_current:
        type: boolean
      last_activity:
        type: string
      project_id:
        type: string
      role_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      user_id_auth:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Token:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      refresh_in_seconds:
        type: integer
      refresh_token:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.V2LoginResponse:
    properties:
      addational_table:
        $ref: '#/definitions/structpb.Struct'
      app_permissions:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.RecordPermission'
        type: array
      client_platform:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.ClientPlatform'
      client_type:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.ClientType'
      environment_id:
        type: string
      global_permission:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GlobalPermission'
      login_table_slug:
        type: string
      permissions:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.RecordPermission'
        type: array
      resource_id:
        type: string
      role:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Role'
      sessions:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Session'
        type: array
      tables:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Object'
        type: array
      token:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Token'
      user:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.User'
      user_data:
        $ref: '#/definitions/structpb.Struct'
      user_found:
        type: boolean
      user_id:
        type: string
      user_id_auth:
        type: string
    type: object
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.V2LoginSuperAdminRes:
    properties:
      companies:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Company'
        type: array
      sessions:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Session'
        type: array
      token:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Token'
      user_data:
        $ref: '#/definitions/structpb.Struct'
      user_found:
        type: boolean
      user_id:
        type: string
    type: object
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.V2LoginWithOptionRequest:
    properties:
      client_id:
        type: string
      client_ip:
        type: string
      data:
        additionalProperties:
          type: string
        type: object
      login_strategy:
        type: string
      tables:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Object'
        type: array
      user_agent:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_object_builder_service.V2LoginResponse:
    properties:
      addational_table:
        $ref: '#/definitions/structpb.Struct'
      app_permissions:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_object_builder_service.RecordPermission'
        type: array
      client_platform:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_object_builder_service.ClientPlatform'
      client_type:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_object_builder_service.ClientType'
      global_permission:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_object_builder_service.GlobalPermission'
      login_table_slug:
        type: string
      permissions:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_object_builder_service.RecordPermission'
        type: array
      role:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_object_builder_service.Role'
      user_data:
        $ref: '#/definitions/structpb.Struct'
      user_found:
        type: boolean
      user_id:
        type: string
      user_id_auth:
        type: string
    type: object
  structpb.Struct:
    properties:
      fields:
        additionalProperties:
          $ref: '#/definitions/structpb.Value'
        description: Unordered map of dynamically typed values.
        type: object
    type: object
...
info:
  contact: {}
  description: This is a api gateway
  termsOfService: https://u-code.io
paths:
...
  /v2/auth/login/{provider}:
    post:
      consumes:
      - application/json
      description: |-
        V2LoginProvider
        inside the data you must be passed client_type_id field
        you must be give environment_id and project_id in body or
        Environment-Id hearder and project-id in query parameters or
        X-API-KEY in hearder
        login strategy must be one of the following values
        ["EMAIL", "PHONE", "EMAIL_OTP", "PHONE_OTP", "LOGIN", "LOGIN_PWD", "GOOGLE_AUTH", "APPLE_AUTH", "PHONE_PWD", "EMAIL_PWD"]
      operationId: V2LoginProvider
      parameters:
      - description: Environment-Id
        in: header
        name: Environment-Id
        type: string
      - description: X-API-KEY
        in: header
        name: X-API-KEY
        type: string
      - description: project-id
        in: query
        name: project-id
        type: string
      - description: V2LoginRequest
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.V2LoginWithOptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.V2LoginSuperAdminRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: V2LoginProvider
      tags:
      - v2_auth
  /v2/auth/logout:
    post:
      consumes:
      - application/json
      description: V2Logout User
      operationId: v2_logout
      parameters:
      - description: LogoutRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.LogoutRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: V2Logout User
      tags:
      - v2_auth
  /v2/auth/verify/{verify_id}:
    post:
      consumes:
      - application/json
      description: V2VerifyProvider
      operationId: V2VerifyProvider
      parameters:
      - description: verify_id
        in: path
        name: verify_id
        required: true
        type: string
      - description: X-API-KEY
        in: header
        name: X-API-KEY
        type: string
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: verify_body
        in: body
        name: verifyBody
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_models.Verify'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.V2LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Verify Otp
      tags:
      - v2_auth
...
  /v2/register:
    post:
      consumes:
      - application/json
      description: |-
        V2Register
        in data must be have type, type must be one of the following values
        ["google", "apple", "email", "phone"]
        client_type_id and role_id must be in body parameters
        you must be give environment_id and project_id in body or
        Environment-Id hearder and project-id in query parameters or
        X-API-KEY in hearder
      operationId: V2register
      parameters:
      - description: X-API-KEY
        in: header
        name: X-API-KEY
        type: string
      - description: Resource-Id
        in: header
        name: Resource-Id
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        type: string
      - description: project-id
        in: query
        name: project-id
        type: string
      - description: register_body
        in: body
        name: registerBody
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_models.RegisterOtp'
      produces:
      - application/json
      responses:
        "201":
          description: Register Data data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.V2LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: V2Register
      tags:
      - v2_register
  /v2/register/{provider}:
    post:
      consumes:
      - application/json
      description: |-
        V2RegisterProvider
        in data must be have type, type must be one of the following values
        ["google", "apple", "email", "phone"]
        client_type_id and role_id must be in body parameters
        you must be give environment_id and project_id in body or
        Environment-Id hearder and project-id in query parameters or
        X-API-KEY in hearder
      operationId: V2RegisterProvider
      parameters:
      - description: provider
        in: path
        name: provider
        required: true
        type: string
      - description: X-API-KEY
        in: header
        name: X-API-KEY
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        type: string
      - description: project-id
        in: query
        name: project-id
        type: string
      - description: register_body
        in: body
        name: registerBody
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_models.RegisterOtp'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.V2LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: V2RegisterProvider
      tags:
      - v2_auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"